+++ projects/modals/src/modules/modal/modal.component.ts
@@ -1,13 +1,18 @@
 import {
+  AfterViewInit,
   Component,
   Input,
   ElementRef,
-  AfterViewInit,
-  HostListener
+  Host,
+  HostListener,
+  ViewChild
 } from '@angular/core';
 
 import {
-  SkyAppWindowRef
+  SkyAppWindowRef,
+  SkyCoreAdapterService,
+  SkyDockLocation,
+  SkyDockService
 } from '@skyux/core';
 
 import {
@@ -46,7 +51,8 @@ let skyModalUniqueIdentifier: number = 0;
     skyAnimationModalState
   ],
   providers: [
-    SkyModalComponentAdapterService
+    SkyModalComponentAdapterService,
+    SkyDockService
   ]
 })
 export class SkyModalComponent implements AfterViewInit {
@@ -114,12 +120,17 @@ export class SkyModalComponent implements AfterViewInit {
 
   public scrollShadow: SkyModalScrollShadowEventArgs;
 
+  @ViewChild('modalContentWrapper', { read: ElementRef })
+  private modalContentWrapperElement: ElementRef;
+
   constructor(
     private hostService: SkyModalHostService,
     private config: SkyModalConfiguration,
     private elRef: ElementRef,
     private windowRef: SkyAppWindowRef,
-    private componentAdapter: SkyModalComponentAdapterService
+    private componentAdapter: SkyModalComponentAdapterService,
+    private coreAdapter: SkyCoreAdapterService,
+    @Host() private dockService: SkyDockService
   ) { }
 
   @HostListener('document:keyup', ['$event'])
@@ -147,7 +158,7 @@ export class SkyModalComponent implements AfterViewInit {
         if (event.which === 9) { // Tab pressed
           let focusChanged = false;
 
-          let focusElementList = this.componentAdapter.loadFocusElementList(this.elRef);
+          let focusElementList = this.coreAdapter.getFocusableChildren(this.elRef.nativeElement);
 
           if (
             event.shiftKey &&
@@ -177,6 +188,12 @@ export class SkyModalComponent implements AfterViewInit {
     // https://stackoverflow.com/questions/40562845
     this.windowRef.nativeWindow.setTimeout(() => {
       this.componentAdapter.modalOpened(this.elRef);
+    });
+
+    this.dockService.setDockOptions({
+      location: SkyDockLocation.ElementBottom,
+      referenceEl: this.modalContentWrapperElement.nativeElement,
+      zIndex: 5
     });
   }